refactored lengthy components and trades reducer
