+++ client/components/stock-listing.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react'
-import {List, Header, Grid} from 'semantic-ui-react'
-import DollarComp from './dollar-comp'
+import {List, Header} from 'semantic-ui-react'
+import StockListingItem from './stock-listing-item'
 
 export default class StockListing extends Component {
   render() {
@@ -16,17 +16,12 @@ export default class StockListing extends Component {
               if (stock.open < stock.latestPrice) diff = 'green'
               if (stock.open > stock.latestPrice) diff = 'red'
               return (
-                <List.Item key={stock.symbol}>
-                  <Grid>
-                    <Grid.Column width={8}>
-                      <span className={diff}>{stock.symbol}</span> -{' '}
-                      {stock.shares} Shares
-                    </Grid.Column>
-                    <Grid.Column width={8} textAlign="right">
-                      <DollarComp money={currentVal} />
-                    </Grid.Column>
-                  </Grid>
-                </List.Item>
+                <StockListingItem
+                  key={stock.symbol}
+                  diff={diff}
+                  currentVal={currentVal}
+                  stock={stock}
+                />
               )
             })}
           </List>
